test-requirements.txt
pytest-asyncio

# for the static typechecker
mypy
